+++ src/main/java/com/kolloseum/fourpillars/interfaces/controller/ContentController.java
@@ -1,25 +1,31 @@
 package com.kolloseum.fourpillars.interfaces.controller;
 
 import com.kolloseum.fourpillars.application.service.FortuneService;
+import com.kolloseum.fourpillars.application.service.NoticeService;
 import com.kolloseum.fourpillars.common.response.ApiResponse;
 import com.kolloseum.fourpillars.common.utils.SecurityContextTokenExtractor;
 import com.kolloseum.fourpillars.interfaces.dto.response.FortuneResponse;
 import com.kolloseum.fourpillars.interfaces.dto.response.MonthlyFortuneResponse;
+import com.kolloseum.fourpillars.interfaces.dto.response.NoticeDetailResponse;
+import com.kolloseum.fourpillars.interfaces.dto.response.NoticeWrapper;
 import com.kolloseum.fourpillars.interfaces.mapper.ResponseMapper;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/contents/fortune")
+@RequestMapping("/contents")
 @RequiredArgsConstructor
 public class ContentController {
 
     private final FortuneService fortuneService;
+    private final NoticeService noticeService;
 
-    @GetMapping("/yearly")
+    @GetMapping("/fortune/yearly")
     public ResponseEntity<ApiResponse<FortuneResponse>> getFortune() {
         var tokenPayload = SecurityContextTokenExtractor.getCurrentTokenPayload();
         var query = fortuneService.getUserFortune(tokenPayload);
@@ -28,12 +34,24 @@ public class ContentController {
         return ResponseEntity.ok(ApiResponse.success(response));
     }
 
-    @GetMapping("/monthly")
+    @GetMapping("/fortune/monthly")
     public ResponseEntity<ApiResponse<MonthlyFortuneResponse>> getMonthlyFortune() {
         var tokenPayload = SecurityContextTokenExtractor.getCurrentTokenPayload();
         var query = fortuneService.getUserMonthlyFortune(tokenPayload);
         var response = ResponseMapper.toMonthlyFortuneResponse(query);
 
+        return ResponseEntity.ok(ApiResponse.success(response));
+    }
+
+    @GetMapping("/notices")
+    public ResponseEntity<ApiResponse<NoticeWrapper>> getNoticeList(Pageable pageable) {
+        var response = noticeService.getNoticeList(pageable);
+        return ResponseEntity.ok(ApiResponse.success(response));
+    }
+
+    @GetMapping("/notices/{id}")
+    public ResponseEntity<ApiResponse<NoticeDetailResponse>> getNoticeDetail(@PathVariable Long id) {
+        var response = noticeService.getNoticeDetail(id);
         return ResponseEntity.ok(ApiResponse.success(response));
     }